Fixed a bunch of problems (warnings and errors) compiling with VC8 for 64 bit.

+++ sql/src/include/sql_mem.h
@@ -79,7 +79,7 @@ extern sql_allocator *sa_create(void);
 extern sql_allocator *sa_reset( sql_allocator *sa );
 extern char *sa_alloc( sql_allocator *sa,  size_t sz );
 extern void sa_destroy( sql_allocator *sa );
-extern char *sa_strndup( sql_allocator *sa, const char *s, int l);
+extern char *sa_strndup( sql_allocator *sa, const char *s, size_t l);
 extern char *sa_strdup( sql_allocator *sa, const char *s);
 
 #define SA_NEW( sa, type ) ((type*)sa_alloc( sa, sizeof(type)) )
